refactor: 왼쪽 사이드바 버튼 컴포넌트화 완료
